refactor: page folder structure
